Manual PCX request acceptance not required.
The documentation currently states "The owner of the accepter VPC must accept the peering request to activate the peering connection. The VPC peering connection request expires after 7 days, after which it cannot be accepted or rejected." but this is not the case when deploying via CloudFormation as far as I can tell.

In order for the stack update/resource creation to complete, the VPC Peering Connection must be created/accepted in both accounts in a single stack update as detailed [here](https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/peer-with-vpc-in-another-account.html) - this is achieved by the requester account assuming a role with adequate permissions in the accepter/peer account.

The proposed changes are an attempt to make this behaviour more clear and not require the reader to do further research to understand this.